Move layouts/shared to containers/layouts
