+++ frontend/containers/layouts/language-selector/types.ts
