+++ frontend/containers/layouts/logo/component.tsx
